requirements.txt
causal_conv1d==1.4.0
einops==0.8.1
flash_attn==2.7.4.post1
h5py==3.11.0
huggingface_hub==0.26.2
lm_eval==0.4.8
mamba_ssm==2.2.2
matplotlib==3.7.5
mne==1.6.1
numpy==1.24.3
opencv_python==4.10.0.84
packaging==25.0
pandas==2.0.3
Pillow==11.2.1
pytest==8.3.5
scikit_learn==1.3.2
scipy==1.15.3
setuptools==75.1.0
torchaudio==2.0.2+cu118
torchsummary==1.5.1
torchvision==0.15.2+cu118
transformers==4.46.3
wheel==0.45.1
